+++ absir-core/src/main/java/com/absir/validator/value/Max.java
@@ -16,5 +16,5 @@ import java.lang.annotation.Target;
 @Retention(RetentionPolicy.RUNTIME)
 public @interface Max {
 
-    int value();
+    float value();
 }
