+++ absir-core/src/main/java/com/absir/validator/value/Min.java
@@ -16,5 +16,5 @@ import java.lang.annotation.Target;
 @Retention(RetentionPolicy.RUNTIME)
 public @interface Min {
 
-    int value();
+    float value();
 }
